per branch files, project-dependant autogen, sanity-check options
work files now comprise the branch name so that multiple branch could
coexist and the build-state is not confused if we switch branch

now autogen.lastrun is pulled from ~/.tinbuild/autogen/<project>.autogen
if it exist... otherwise we use the curretn autogen.lastrun

added -c switch to trigger a 'sanity check'. this verify that the config
fiels are where they are supposed to be and contain the minimum
necessary to send an email. a test email is sent to the owner.