allow for the build history info to be in a separate directory
so far info relative to the last sucessfull build, the last upload etc.
where store directly in the root the source tree.
in order to allow for the whole thing to be in
a ramdisk, and still not have to re-prime each time the ramdisk
as to be re-created it was needed to store these information
somewhere else.
now the config file can specify METADATA_DIR to that effect
by default this is '.' which is backward comaptible.