pull again after having waited on a lock
Otherwise, with 2 tinderboxes on the same machine, each of them needing
e.g. 1 hour to build, the first one pulls and builds, the second one pulls,
waits 1 hour for the first one, and then it starts the build, with its
checkout already 1 hour out of date; and then they both alternate this way.